# docs/api-contract/api-contract.booking-service.yaml
openapi: 3.0.3

info:
  title: booking-service API
  version: 1.0.0
  description: >
    booking-service owns the Service, TimeSlot and Appointment entities for the
    Dog Grooming Appointment Booking System.

    Scope of this revision: PRD F1 (the public Service list consumed by Screen 1),
    F2/F3/F3b — the public TimeSlot availability query and the atomic hold
    consumed by Screen 2 (Time Slot Picker) — and F4/F4b, Appointment creation
    from the Customer Details form (Screen 3). The Admin Service writes (F6-F8)
    and the Admin Appointment management routes (F9-F11) are defined by their own
    tickets and are deliberately absent here.

    Every route below is public and unauthenticated by design.

servers:
  - url: http://localhost:4001
    description: Local development

tags:
  - name: Service
    description: Treatments the clinic offers.
  - name: TimeSlot
    description: Bookable windows for a Service on a given day.
  - name: Appointment
    description: A Customer's booking of a Service at a TimeSlot.

paths:
  /api/services:
    get:
      tags: [Service]
      operationId: listActiveServices
      summary: List all active Services
      description: >
        Public and unauthenticated by design (PRD: "all public booking routes
        are unauthenticated"). Returns ONLY Services with `isActive: true` —
        a deactivated Service is soft-deleted and must never appear on the
        customer-facing list (PRD AC-4).

        This endpoint accepts no query parameters. Any query string sent by a
        client must be ignored outright and must never be merged into the
        database filter.
      security: []
      responses:
        '200':
          description: The active Services. An empty array is a valid, expected result.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
              examples:
                populated:
                  summary: Two active Services
                  value:
                    - id: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
                      name: Full groom — small dog
                      durationMinutes: 90
                      price: 220
                      isActive: true
                    - id: 1a2b3c4d-5e6f-4071-8293-a4b5c6d7e8f9
                      name: Bath and blow dry
                      durationMinutes: 45
                      price: 120
                      isActive: true
                empty:
                  summary: No active Services
                  value: []
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: The database is unreachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/time-slots:
    get:
      tags: [TimeSlot]
      operationId: listOpenTimeSlots
      summary: List the bookable TimeSlots for one Service on one day
      description: >
        Public and unauthenticated (PRD F2). Returns ONLY slots whose effective
        status is `open` for that Service and calendar day. A slot that is
        `held` or `booked` must never appear in this response — the frontend
        filters defensively, but it treats every returned slot as claimable.

        Lazy hold expiry (PRD F3b): a `held` slot whose `heldAt` is older than
        the hold TTL has effectively expired and MUST be returned here as
        available, with `status: "open"`. There is no background sweeper yet, so
        this read is what heals a stale hold.

        Both query parameters are required. Reject a missing or malformed
        `date` (anything that is not a real `YYYY-MM-DD` day) or an invalid
        `serviceId` with 400 rather than coercing it — and never merge raw query
        input into the database filter.
      security: []
      parameters:
        - name: serviceId
          in: query
          required: true
          description: The Service whose availability is being browsed.
          schema:
            type: string
            format: uuid
          example: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
        - name: date
          in: query
          required: true
          description: >
            The calendar day, in the clinic's local timezone. A plain calendar
            day, never a UTC instant — the clinic is single-location and the
            wall-clock day is what a customer picks.
          schema:
            type: string
            format: date
            pattern: '^\d{4}-\d{2}-\d{2}$'
          example: '2026-08-18'
      responses:
        '200':
          description: >
            The bookable slots for that Service and day. An empty array is a
            valid, expected result — a fully booked day is not an error.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
              examples:
                populated:
                  summary: Two bookable slots
                  value:
                    - id: 3c9d1e2f-4a5b-4c6d-8e7f-9a0b1c2d3e4f
                      serviceId: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
                      date: '2026-08-18'
                      startTime: '09:00'
                      endTime: '10:30'
                      status: open
                    - id: 5e7f9a0b-1c2d-4e3f-8a4b-5c6d7e8f9a0b
                      serviceId: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
                      date: '2026-08-18'
                      startTime: '11:00'
                      endTime: '12:30'
                      status: open
                fullyBooked:
                  summary: Nothing left that day
                  value: []
        '400':
          description: A missing/malformed `serviceId` or `date`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: The database is unreachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/time-slots/{id}/hold:
    post:
      tags: [TimeSlot]
      operationId: holdTimeSlot
      summary: Atomically claim an open TimeSlot
      description: >
        Public and unauthenticated (PRD F3). Transitions the slot `open` →
        `held` and stamps `heldAt`, so the customer can complete Screen 3
        without losing the time to someone else.

        This MUST be a single atomic conditional update — e.g.
        `findOneAndUpdate({ _id, status: 'open' }, { $set: { status: 'held',
        heldAt: now } })` — so the datastore itself enforces exclusivity. A
        read-then-write is a race and is not acceptable here: two simultaneous
        requests for the same slot must resolve to exactly one 200 and one 409,
        never two 200s (see `seat-concurrency-layer`).

        A slot whose hold has already expired past the TTL counts as `open` for
        this transition and may be re-held.

        The response MUST include `holdExpiresAt` — `heldAt` plus the hold TTL,
        as an absolute instant. Screen 3 shows the Customer how long is left so
        the form does not lapse silently under them. This is a display value
        only: the server never trusts it back, and `POST /api/appointments`
        re-checks expiry itself. See the `holdExpiresAt` note on the TimeSlot
        schema for why the deadline is exposed but `heldAt` is not.

        The request has no body. This endpoint mutates state without auth by
        design; it is a deliberate abuse surface, and rate limiting is a known
        follow-up rather than something silently assumed here.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TimeSlot being claimed.
          schema:
            type: string
            format: uuid
          example: 3c9d1e2f-4a5b-4c6d-8e7f-9a0b1c2d3e4f
      responses:
        '200':
          description: >
            The hold succeeded. Returns the updated slot, which the client
            treats as the only proof the hold landed — it never assumes success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'
              example:
                id: 3c9d1e2f-4a5b-4c6d-8e7f-9a0b1c2d3e4f
                serviceId: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
                date: '2026-08-18'
                startTime: '09:00'
                endTime: '10:30'
                status: held
                holdExpiresAt: '2026-08-18T09:05:00.000Z'
        '400':
          description: The `id` is not a well-formed identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No TimeSlot with that id exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >
            The conditional update matched zero documents: the slot is already
            `held` (unexpired) or `booked`. This is an expected outcome of
            normal concurrent use, not a fault — the frontend maps it to a
            specific "that time was just taken" message and re-fetches the list,
            distinct from how it treats any other failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: TimeSlot is no longer open
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: The database is unreachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/appointments:
    post:
      tags: [Appointment]
      operationId: createAppointment
      summary: Book a held TimeSlot into an Appointment
      description: >
        Public and unauthenticated (PRD F4) — a Customer has no account and
        never logs in. This is the second half of the hold→book lifecycle whose
        first half is `POST /api/time-slots/{id}/hold`.

        The server MUST, in order:

        1. Validate the body. `customerName` and `customerPhone` are required;
           `customerEmail` is optional but must be a well-formed address when
           present. Reject with 400 — never coerce, never store a partial
           record. `customerName` is attacker-controlled free text: store it
           verbatim, and never interpolate it into a query or a notification
           template without escaping.
        2. Load the TimeSlot named by `slotId`, and confirm it is currently
           `held` and its hold has NOT lapsed past the TTL, and that its
           `serviceId` matches the body's. A lapsed hold counts as `open`
           (PRD F3b) and must NOT be bookable — that slot is available to
           everyone again.
        3. Transition the slot `held` → `booked` with a SINGLE atomic
           conditional update, exactly as the hold does — e.g.
           `findOneAndUpdate({ _id, status: 'held' }, { $set: { status:
           'booked' } })`. A read-then-write is a race and is not acceptable:
           two simultaneous submits for the same held slot (a double-tap, a
           retried request) must resolve to exactly one 201 and one 409, never
           two Appointments for one slot (see `seat-concurrency-layer`).
        4. Create the Appointment with `status: 'pending'`. It is never created
           `confirmed`: confirming is the Admin's decision (PRD F9-F11).
        5. Fire a best-effort call to notification-service's confirmation
           endpoint (PRD F4b). This MUST NOT block or fail this response:
           catch and log any error from it. A notification-service outage must
           never cost a Customer their booking, and must never roll one back
           that the datastore has already accepted.

        The client treats this response as the only proof the booking landed —
        it never assumes success, exactly as with the hold.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
            example:
              slotId: 3c9d1e2f-4a5b-4c6d-8e7f-9a0b1c2d3e4f
              serviceId: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
              customerName: Dana Levi
              customerPhone: 050-123-4567
              customerEmail: dana@example.com
      responses:
        '201':
          description: >
            The Appointment was created and the TimeSlot is now `booked`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              example:
                id: 7b2e4d6f-8a91-4c3b-9d5e-1f2a3b4c5d6e
                serviceId: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
                timeSlotId: 3c9d1e2f-4a5b-4c6d-8e7f-9a0b1c2d3e4f
                customerName: Dana Levi
                customerPhone: 050-123-4567
                customerEmail: dana@example.com
                status: pending
        '400':
          description: >
            A missing/malformed contact field, a malformed `slotId`, or a
            `serviceId` that does not match the slot's own.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No TimeSlot with that `slotId` exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >
            The conditional update matched zero documents: the hold had lapsed,
            the slot was never `held`, or another request booked it first. No
            Appointment was created and the TimeSlot was not mutated.

            This is an expected outcome, not a fault — holds are short-lived by
            design, so a Customer who lingers on the form legitimately arrives
            too late. The frontend maps it to a specific "the hold on this time
            ran out" explanation and a route back to the Time Slot Picker,
            distinct from how it treats any other failure. It is also the
            authoritative backstop for the client's countdown, which is advisory
            only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: TimeSlot is no longer held
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: The database is unreachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Service:
      type: object
      description: >
        A treatment the clinic offers. The client-facing identifier is always
        `id`, the record's uuid — Mongo's `_id` and `__v` are internal and must
        be stripped by the model's toJSON transform, never serialized here
        (.rule/database-rules.md, "External Identity").
      required: [id, name, durationMinutes, price, isActive]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
          description: The Service's uuid. Never a Mongo ObjectId.
          example: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
        name:
          type: string
          minLength: 1
          example: Full groom — small dog
        durationMinutes:
          type: integer
          minimum: 1
          description: Length of the treatment in minutes.
          example: 90
        price:
          type: number
          minimum: 0
          description: Price in ILS. The clinic is single-currency in v1.
          example: 220
        isActive:
          type: boolean
          description: >
            Soft-delete flag. Always `true` in this endpoint's response, since
            it filters on it; included so the same schema serves the Admin list
            (F6-F8), which does return inactive records.
          example: true

    TimeSlot:
      type: object
      description: >
        One bookable window for a Service on a given day. As with Service, the
        client-facing identifier is always `id`, the record's uuid — Mongo's
        `_id` and `__v` are internal and must be stripped by the model's toJSON
        transform (.rule/database-rules.md, "External Identity").

        `heldAt` is deliberately NOT part of this payload: when a hold *started*
        is server-side bookkeeping, and exposing it would invite a client to
        recompute an expiry only the server can decide. The already-computed
        deadline, `holdExpiresAt`, is exposed instead — see that property.
      required: [id, serviceId, date, startTime, endTime, status]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
          description: The TimeSlot's uuid. Never a Mongo ObjectId.
          example: 3c9d1e2f-4a5b-4c6d-8e7f-9a0b1c2d3e4f
        serviceId:
          type: string
          format: uuid
          description: The Service this window is bookable for.
          example: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
        date:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: >
            The calendar day in the clinic's local timezone. A plain day string,
            never a UTC instant, so the day can never shift under a client.
          example: '2026-08-18'
        startTime:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          description: >
            Wall-clock start in 24h `HH:mm`. Carries no timezone by design — the
            client renders it verbatim rather than re-parsing it into a Date.
          example: '09:00'
        endTime:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          description: Wall-clock end in 24h `HH:mm`.
          example: '10:30'
        status:
          type: string
          enum: [open, held, booked]
          description: >
            `open` — bookable by anyone.
            `held` — transiently claimed by a customer mid-checkout; expires
            server-side back to `open` after the hold TTL.
            `booked` — finalized into an Appointment.

            The public list endpoint only ever returns `open`; `held` is what a
            successful hold responds with.
          example: open
        holdExpiresAt:
          type: string
          format: date-time
          description: >
            When this hold lapses, as an absolute ISO-8601 instant (`heldAt` +
            the hold TTL, computed server-side).

            Present ONLY on the response to a successful hold, where `status` is
            `held`; omitted everywhere else, including every entry in the public
            list. Unlike `startTime`/`endTime`, which are wall-clock strings
            carrying no timezone, this one genuinely is an instant — it is a
            deadline, not a time of day, and has to be comparable against the
            Customer's own clock.

            It exists so Screen 3 can show the Customer how much time is left
            rather than letting the form expire in silence. It is advisory: the
            client's countdown never decides whether a booking may proceed, the
            409 from `POST /api/appointments` does.
          example: '2026-08-18T09:05:00.000Z'

    Appointment:
      type: object
      description: >
        A Customer's booking of a Service at a TimeSlot. As above, the
        client-facing identifier is always `id`, the record's uuid — Mongo's
        `_id` and `__v` are internal and must be stripped by the model's toJSON
        transform (.rule/database-rules.md, "External Identity").

        This record holds personally identifiable contact details. Only the
        fields below may be serialized to a Customer: no internal notes, no
        audit fields, and never another Customer's record.
      required: [id, serviceId, timeSlotId, customerName, customerPhone, status]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
          description: The Appointment's uuid. Never a Mongo ObjectId.
          example: 7b2e4d6f-8a91-4c3b-9d5e-1f2a3b4c5d6e
        serviceId:
          type: string
          format: uuid
          description: >
            The Service booked. Derived server-side from the TimeSlot document,
            never copied from the request body — see the request schema.
          example: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
        timeSlotId:
          type: string
          format: uuid
          description: The TimeSlot this Appointment finalized to `booked`.
          example: 3c9d1e2f-4a5b-4c6d-8e7f-9a0b1c2d3e4f
        customerName:
          type: string
          minLength: 1
          maxLength: 60
          description: >
            Free text, and the one field here a Customer fully controls. Store
            and return it verbatim; it must never be interpolated into a query,
            a template or a notification body without escaping.
          example: Dana Levi
        customerPhone:
          type: string
          minLength: 9
          maxLength: 20
          description: >
            The Customer's contact number, as typed. Formatting characters
            (spaces, hyphens, parentheses, a leading `+`) are permitted — the
            clinic dials this, it is not a normalized key.
          example: 050-123-4567
        customerEmail:
          type: string
          format: email
          maxLength: 254
          description: >
            Optional by design (PRD F4): a Customer without email can still
            book. Absent — not an empty string — when not supplied.
          example: dana@example.com
        status:
          type: string
          enum: [pending, confirmed, cancelled]
          description: >
            `pending`   — created by a Customer, awaiting the Admin's decision.
            `confirmed` — the Admin approved it.
            `cancelled` — the Admin cancelled it.

            This endpoint always creates `pending`. Moving to `confirmed` or
            `cancelled` is the Admin's alone (PRD F9-F11) and belongs to the
            Admin routes, which are not in this revision.
          example: pending

    CreateAppointmentRequest:
      type: object
      description: >
        What Screen 3 submits. `slotId` is the source of truth: the server MUST
        resolve the Service, date and time from that TimeSlot document rather
        than from anything else in this body, so a client cannot book one
        Service's slot under another's name.
      required: [slotId, serviceId, customerName, customerPhone]
      additionalProperties: false
      properties:
        slotId:
          type: string
          format: uuid
          description: The TimeSlot this Customer currently holds.
          example: 3c9d1e2f-4a5b-4c6d-8e7f-9a0b1c2d3e4f
        serviceId:
          type: string
          format: uuid
          description: >
            The Service the client believes it is booking. Sent for routing and
            as a cross-check only — the server MUST verify it matches the slot's
            own `serviceId` and reject a mismatch with 400, never trust it over
            the slot document.
          example: 8f1c2b34-5d6e-4f70-8a91-2b3c4d5e6f70
        customerName:
          type: string
          minLength: 1
          maxLength: 60
          example: Dana Levi
        customerPhone:
          type: string
          minLength: 9
          maxLength: 20
          example: 050-123-4567
        customerEmail:
          type: string
          format: email
          maxLength: 254
          description: >
            Optional. The client omits the field entirely when the Customer
            leaves it blank rather than sending an empty string, so absence is
            unambiguous.
          example: dana@example.com

    Error:
      type: object
      description: >
        Uniform error envelope. The message is for logs and developers — the
        frontend never renders it, it maps the status code to its own localized
        copy (.rule/error-handling-rules.md).
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
          example: Internal Server Error